Move `ClientConnection` class to its own file
